The veins can be loaded now (but not displayed).

+++ NeckVeins 1.0/src/main/Camera.java
@@ -6,9 +6,9 @@ public class Camera {
 	public static final float CAMERA_MOVE_SPEED = 1.667f;
 	public static final double CAMERA_ROTATION_SPEED = (float) (72 * Math.PI / 180 / 60);
 
-	public int cameraX;
-	public int cameraY;
-	public int cameraZ;
+	public float cameraX;
+	public float cameraY;
+	public float cameraZ;
 	public Quaternion cameraOrientation;
 
 	public Camera() {
